requirements.txt
fastapi
uvicorn
telethon
sqlite3
python-dotenv
cryptography